refactor(user): 컴포넌트 CurrentScheduleBox 이름 변경
