+++ apps/user/src/components/main/ApplicationSubmittionBox/ApplicationPeriodBox/ApplicationPeriodBox.hooks.ts
