+++ apps/user/src/components/main/ApplicationSubmittionBox/ApplicationSubmittionBox.hooks.ts
