+++ apps/user/src/components/main/index.ts
@@ -1,5 +1,5 @@
 export { default as ApplicationBox } from './ApplicationBox/ApplicationBox';
-export { default as CurrentScheduleBox } from './CurrentScheduleBox/CurrentScheduleBox';
+export { default as ApplicationSubmittionBox } from './ApplicationSubmittionBox/ApplicationSubmittionBox';
 export { default as FaqBox } from './FaqBox/FaqBox';
 export { default as GuidelineBox } from './GuidelineBox/GuidelineBox';
 export { default as NoticeBox } from './NoticeBox/NoticeBox';
